ui/rbac-poc-actions-menu: new edit menu button
